redirect only the target argument of mount
The source is usually starts with /dev/ and is not need to be
redirected. The result of redirecting it will look like
/./../././dev/sda2
in an output of mount or df, which is bad.
If someone in the future wants to redirect the source of bind mounts,
the new special wrapper should be added that will check for MS_BIND flag.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>