remove AC_FUNC_MALLOC
This checks for the gnu malloc, that is, the one that returns
non-NULL on malloc(0). If such malloc is not found, it defines
malloc to rpl_malloc, which is in gnulib. You get an undefined
reference as a result.
http://www.uclibc.org/FAQ.html#gnu_malloc

Since user-union doesn't use malloc(0) (no sane code does), it
is safe to just remove AC_FUNC_MALLOC.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>