wrap all functions that are not specified to be signal-safe
This patch creates the private libc namespaces and wraps all
the functions that are not specified to be signal-safe by posix.

As you can see from the following output, some libdl functions are
still linked in, and dlsym() is used on a critical path. I really
hope dlsym() is signal-safe because the lazy symbol binding can
happen at any time.

$ nm -S user-union.o |grep " U "
                 U _GLOBAL_OFFSET_TABLE_
                 U __assert_fail
                 U __errno_location
                 U _exit
                 U abort
                 U close
                 U create_branchlist
                 U dlclose
                 U dlmopen
                 U dlsym
                 U read
                 U stderr
                 U write

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>