implement unlink() atomically
With the added functionality it is now possible to implement unlink()
without creating a temporary file.
This fixes a horrible data corruption: if the process is killed
after it created an empty file in overlay, on the next start you'll
still have that empty file as overlay for the real file!

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>